pyproject.toml
[tool.poetry]
name = "mofa"
version = "0.1.0"
description = ""
authors = ["Jakob Duerrwaechter <jakob.duerrwaechter@pik-potsdam.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
matplotlib = "^3.7.1"
numpy = "^1.25.0"
scipy = "^1.11.0"
PyYAML = "^6.0"
statsmodels = "^0.14.0"
pandas = "^2.0.2"
pickle4 = "^0.0.1"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
xlwt = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
